requirements.txt
# MCP Core Dependencies
mcp>=1.0.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
PyMuPDF>=1.23.0

# LangChain for Advanced Text Processing & Chunking
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# Text Processing & Search
sentence-transformers>=2.2.0
chromadb>=0.4.22
transformers>=4.21.0

# File System & Obsidian Integration
watchdog>=3.0.0
python-frontmatter>=1.0.0
markdown>=3.5.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.0.0

# Utility Libraries
requests>=2.31.0
pathlib>=1.0.0
regex>=2023.0.0
tiktoken>=0.5.0 